+++ ChangeLog.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 
 ## 1.1
 
+- FIX : Traduction sur facture de situation *26/06/2021* - 1.1.8
 - FIX : module descriptor (v14 compatibility) *30/06/2021* - 1.1.7
 - FIX : Remove useless dependency for workstation *29/06/2021* - 1.1.6
 - FIX : V13 GETPOST compatibility *08/03/2021*